requirements.txt
# Core dependencies
websockets>=12.0
python-dotenv>=1.0.0
